+++ tests/test_formparser.py
@@ -10,6 +10,8 @@
 """
 from __future__ import with_statement
 
+import csv
+import io
 import pytest
 
 from os.path import join, dirname
@@ -134,44 +136,32 @@ class TestFormParser(object):
         strict_eq(len(form), 0)
         strict_eq(len(files), 0)
 
-    @pytest.mark.skipif(PY2, reason='io "File Objects" are only available in PY3.')
     @pytest.mark.parametrize(
-        "attributes", (
-            # make sure we have a real file here, because we expect to be
-            # on the disk.  > 1024 * 500
-            "fileno",
-            # Make sure the file object has the bellow attributes as described
-            # in https://github.com/pallets/werkzeug/issues/1344
-            "writable", "readable", "seekable"
-        )
+        ("no_spooled", "size"),
+        ((False, 100), (False, 3000), (True, 100), (True, 3000)),
     )
-    def test_large_file(self, attributes):
-        data = b'x' * (1024 * 600)
-        req = Request.from_values(data={'foo': (BytesIO(data), 'test.txt')},
-                                  method='POST')
-        file_storage = req.files['foo']
-        assert hasattr(file_storage, attributes)
-        # close file to prevent fds from leaking
-        req.files['foo'].close()
+    def test_default_stream_factory(self, no_spooled, size, monkeypatch):
+        if no_spooled:
+            monkeypatch.setattr("werkzeug.formparser.SpooledTemporaryFile", None)
 
-    @pytest.mark.skipif(PY2, reason='io "File Objects" are only available in PY3.')
-    @pytest.mark.parametrize(
-        "attributes", (
-            # Make sure the file object has the bellow attributes as described
-            # in https://github.com/pallets/werkzeug/issues/1344
-            "writable", "readable", "seekable"
+        data = b"a,b,c\n" * size
+        req = Request.from_values(
+            data={"foo": (BytesIO(data), "test.txt")},
+            method="POST"
         )
-    )
-    def test_small_file(self, attributes):
-        # when the data's length is below the "max_size", this will implement
-        # an in-memory buffer
-        data = b'x' * 256
-        req = Request.from_values(data={'foo': (BytesIO(data), 'test.txt')},
-                                  method='POST')
-        file_storage = req.files['foo']
-        assert hasattr(file_storage, attributes)
-        # close file to prevent fds from leaking
-        req.files['foo'].close()
+        file_storage = req.files["foo"]
+
+        try:
+            if PY2:
+                reader = csv.reader(file_storage)
+            else:
+                reader = csv.reader(io.TextIOWrapper(file_storage))
+            # This fails if file_storage doesn't implement IOBase.
+            # https://github.com/pallets/werkzeug/issues/1344
+            # https://github.com/python/cpython/pull/3249
+            assert sum(1 for _ in reader) == size
+        finally:
+            file_storage.close()
 
     def test_streaming_parse(self):
         data = b'x' * (1024 * 600)