+++ werkzeug/datastructures.py
@@ -2758,6 +2758,9 @@ class FileStorage(object):
         try:
             return getattr(self.stream, name)
         except AttributeError:
+            # SpooledTemporaryFile doesn't implement IOBase, get the
+            # attribute from its backing file instead.
+            # https://github.com/python/cpython/pull/3249
             if hasattr(self.stream, "_file"):
                 return getattr(self.stream._file, name)
             raise